[RED] TEST: REFACTOR and REMOVE test for fish_sampling to get latest fish sampling and use datetime
